requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
decorator==4.4.0
ffmpeg-python==0.2.0
Flask==1.1.1
Flask-Caching==1.7.2
future==0.17.1
idna==2.8
imageio==2.6.1
imageio-ffmpeg==0.3.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
moviepy==1.0.0
numpy==1.17.2
oauthlib==3.1.0
Pillow==6.2.0
proglog==0.1.9
python-dateutil==2.8.0
python-twitter==3.5
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
tqdm==4.36.1
urllib3==1.25.6
Werkzeug==0.16.0
